Added domensionality tags for views. They specify which selectors can be triggered on the views. Implemented Range2DInputGroup. Moved input groups to a separate module. Fixed serialization of Range2DParameterGroup
