+++ pymeasure/display/widgets/image_widget.py
@@ -40,11 +40,12 @@ class ImageWidget(TabWidget, QtWidgets.QWidget):
     """
 
     def __init__(self, name, columns, x_axis, y_axis, z_axis=None, refresh_time=0.2,
-                 check_status=True, parent=None):
+                 check_status=True, selectors: list = [], parent=None):
         super().__init__(name, parent)
         self.columns = columns
         self.refresh_time = refresh_time
         self.check_status = check_status
+        self.selectors = selectors
         self.x_axis = x_axis
         self.y_axis = y_axis
         self._setup_ui()
@@ -113,3 +114,19 @@ class ImageWidget(TabWidget, QtWidgets.QWidget):
 
     def remove(self, curve):
         self.plot.removeItem(curve)
+
+    def selection_tags(self) -> list[str]:
+        return ["1D", "2D"]
+        
+    def enter_selection_mode(self, initial_range):
+        super().enter_selection_mode()
+        start_x, stop_x, start_y, stop_y = initial_range
+        self.rect_reg = pg.RectROI([start_x, stop_x], [stop_x-start_x, stop_y-start_y])
+        self.rect_reg.setPen(pg.mkPen("black"))
+        self.image_frame.plot_widget.addItem(self.rect_reg)
+
+    def exit_selection_mode(self):
+        start_x, start_y = self.rect_reg.pos()
+        size_x, size_y = self.rect_reg.size()
+        self.image_frame.plot_widget.removeItem(self.rect_reg)
+        return (start_x, start_x+size_x, start_y, start_y+size_y)
